FlaskMDM/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
google-api-python-client==2.141.0
google-auth==2.33.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
